+++ MinaTolWebApi/DAL/DbWrapper.UnidadMedida.cs
@@ -61,7 +61,7 @@ namespace MinaTolWebApi.DAL
             return response;
         }
 
-        public ModelResponse GetUnidadMedidaById(int id)
+        public ModelResponse GetUnidadMedidaById(long id)
         {
             var response = new ModelResponse();
             var parameters = new List<SqlParameter>();
